test: consolidate contact fixture helpers, drop redundant comments
Folds emailForName and phoneNumberArbitrary into contactFields.ts since
both are only used there/by customers+vendors schema.ts, and strips
comments that just restated what the code does.
